app/Rag/requirements.txt
flask
google-generativeai
chromadb
pypdf
python-dotenv
flask-cors
